Node: carry Binders, not CoreNames, in nOutputs
This change is motivated by a need for type and clock information for
node outputs. Prior to this change, only CoreNames for outputs were
available from the Node type, requiring a user to discover outputs'
types and clocks some other way. Since the type and clock information
was already known but was being discarded, this change just preserves
that information (as Binders) in nOutputs instead. Otherwise this change
should be functionality-preserving.

This change also adds nOutputs to the nodeEnv typing environment for a
node.