Core: make Merge constructor carry more specific information, stop unfolding
Prior to this change, a Merge carried a conditional expression, a
then-expression, and an else-expression. That meant that a Merge from
an input program would be unfolded into a Merge that either evaluated
to its then-branch or had an else-branch that contained the rest of
the Merge's alternatives. That made for a simple implementation in the
Semantics.Core evaluator, but it mean that consumers of Core that needed
the original Merge structure would have a very difficult time recovering
it.

This change modifies Merge so that no unfolding is done, but the
identifier and type of the merge expression is carried, and all of the
alternatives are carried together in a single Merge constructor.